[[ Bug 16029 ]] Missing AS Dictionary
LiveCode commands are missing from the AS Dictionary in builds as far back an 7.1 dp1 and 8.0 dp4. When the AS dictionary is viewed in Script Editor, only Standard Suite, Text Suite, and Type definition are available.

I'm unsure how the information made it into earlier versions, but Apple documentation indicates that the .sdef file should be referenced in the Info.plist file. I've edited the 4 .plist files that enable AS and added the .sdef file reference (unsure about Test-Info.plist - not sure if the link needs to include the folder since it sits one level higher in the directory structure).

Additionally, the .sdef file was updated to include the reference to the Standard Suite of commands [[Bug 8577]]

I have not set up the full build environment, so I have not done a full build. I did edit the Info.plist for my 9.0.0 DP 8 copy and included the terminology.sdef Resources folder in the app package and was able to see the entries. I was also able to use "do script" from SE.

---

Take 2... I did a community build on develop and it worked properly.  I tried to rebase on 8.1 and something went very wrong.